NP-28592 Rename folder
