+++ src/pages/search/SearchPage.tsx
@@ -4,11 +4,11 @@ import { useHistory } from 'react-router-dom';
 import SearchIcon from '@mui/icons-material/Search';
 import ManageSearchIcon from '@mui/icons-material/ManageSearch';
 import { Box, Button } from '@mui/material';
-import { RegistrationSearchBar } from './result_search/RegistrationSearchBar';
+import { RegistrationSearchBar } from './registration_search/RegistrationSearchBar';
 import { createSearchConfigFromSearchParams, createSearchQuery, SearchParam } from '../../utils/searchHelpers';
-import { RegistrationFacetsFilter } from './result_search/filters/RegistrationFacetsFilter';
-import { RegistrationSearch } from './result_search/RegistrationSearch';
-import { RegistrationSortSelector } from './result_search/RegistrationSortSelector';
+import { RegistrationFacetsFilter } from './registration_search/filters/RegistrationFacetsFilter';
+import { RegistrationSearch } from './registration_search/RegistrationSearch';
+import { RegistrationSortSelector } from './registration_search/RegistrationSortSelector';
 import { SearchResponse } from '../../types/common.types';
 import { Registration } from '../../types/registration.types';
 import { useFetch } from '../../utils/hooks/useFetch';
