+++ src/pages/search/registration_search/RegistrationSearchBar.tsx
@@ -4,7 +4,7 @@ import SearchIcon from '@mui/icons-material/Search';
 import FilterAltIcon from '@mui/icons-material/FilterAltOutlined';
 import { Field, FieldArray, FieldArrayRenderProps, FieldProps, useFormikContext } from 'formik';
 import { ExpressionStatement, PropertySearch, SearchConfig } from '../../../utils/searchHelpers';
-import { AdvancedSearchRow } from './filters/AdvancedSearchRow';
+import { AdvancedSearchRow } from '../registration_search/filters/AdvancedSearchRow';
 import { SearchTextField } from '../SearchTextField';
 
 export const RegistrationSearchBar = () => {
