+++ src/pages/search/registration_search/RegistrationSearchResults.tsx
