+++ src/pages/search/registration_search/filters/RegistrationFacetsFilter.tsx
